requirements.txt
pygame==1.9.6
numpy==1.19.5
pandas==1.0.3
matplotlib==3.2.1
opencv_python==4.5.1.48
tensorflow==2.6.0
PyDirectInput==1.0.4
Pillow==8.3.1
